+++ ViewModels/ShoppingCartViewModel.cs
@@ -4,12 +4,12 @@ namespace BethanysPieShop.ViewModels;
 
 public class ShoppingCartViewModel
 {
-    public ShoppingCartViewModel(IShoppingCart shoppingCart, decimal shoppingCartTotal)
+    public ShoppingCartViewModel(IShoppingCart shoppingCart, double shoppingCartTotal)
     {
         ShoppingCart = shoppingCart;
         ShoppingCartTotal = shoppingCartTotal;
     }
     
     public IShoppingCart ShoppingCart { get; }
-    public decimal ShoppingCartTotal { get; }
+    public double ShoppingCartTotal { get; }
 }
